refactor: extract getEventTypesPublic for public event types to book
