+++ packages/lib/getEventTypesPublic.ts
@@ -1,0 +1,72 @@
+import logger from "@calcom/lib/logger";
+import prisma from "@calcom/prisma";
+import { baseEventTypeSelect } from "@calcom/prisma/selects";
+import { EventTypeMetaDataSchema } from "@calcom/prisma/zod-utils";
+
+import { markdownToSafeHTML } from "./markdownToSafeHTML";
+
+const log = logger.getSubLogger({ prefix: ["getEventTypesPublic"] });
+
+export async function getEventTypesPublic(userId: number) {
+  const eventTypesWithHidden = await getEventTypesWithHiddenFromDB(userId);
+
+  const eventTypesRaw = eventTypesWithHidden.filter((evt) => !evt.hidden);
+
+  return eventTypesRaw.map((eventType) => ({
+    ...eventType,
+    metadata: EventTypeMetaDataSchema.parse(eventType.metadata || {}),
+    descriptionAsSafeHTML: markdownToSafeHTML(eventType.description),
+  }));
+}
+
+const getEventTypesWithHiddenFromDB = async (userId: number) => {
+  const eventTypes = await prisma.eventType.findMany({
+    where: {
+      AND: [
+        {
+          teamId: null,
+        },
+        {
+          OR: [
+            {
+              userId,
+            },
+            {
+              users: {
+                some: {
+                  id: userId,
+                },
+              },
+            },
+          ],
+        },
+      ],
+    },
+    orderBy: [
+      {
+        position: "desc",
+      },
+      {
+        id: "asc",
+      },
+    ],
+    select: {
+      ...baseEventTypeSelect,
+      metadata: true,
+    },
+  });
+  // map and filter metadata, exclude eventType entirely when faulty metadata is found.
+  // report error to exception so we don't lose the error.
+  return eventTypes.reduce<typeof eventTypes>((eventTypes, eventType) => {
+    const parsedMetadata = EventTypeMetaDataSchema.safeParse(eventType.metadata);
+    if (!parsedMetadata.success) {
+      log.error(parsedMetadata.error);
+      return eventTypes;
+    }
+    eventTypes.push({
+      ...eventType,
+      metadata: parsedMetadata.data,
+    });
+    return eventTypes;
+  }, []);
+};
