+++ packages/lib/index.ts
@@ -6,3 +6,4 @@ export * from "./isDurationLimits";
 export * from "./validateIntervalLimitOrder";
 export * from "./schedules";
 export * from "./getEventTypeByViewer";
+export * from "./getEventTypesPublic";
